# swagger.yaml
openapi: 3.0.3
info:
  title: Retail POS Microservices API
  description: API documentation for the Retail Point of Sale (POS) backend microservices with AI-observable telemetry, multi-lane context headers, and chaos injection endpoints.
  version: 1.2.0
servers:
  - url: http://localhost:8081
    description: Order Service
  - url: http://localhost:8083
    description: Payment Service
  - url: http://localhost:8084
    description: Finalize Service
  - url: http://localhost:8085
    description: Telemetry Service
...
paths:
...
  /simulation/failures:
    post:
      summary: Chaos & Failure Injection Endpoint
      description: Inject synthetic timeouts or database failures for Dev Mode AI Ops testing (Available on ports 8081, 8083, 8084).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - type:
                - enabled:
              properties:
                type:
                  type: string
                  example: PAYMENT_TIMEOUT
                enabled:
                  type: boolean
                  example: true
                duration_ms:
                  type: integer
                  example: 5000
      responses:
        '200':
          description: Simulation failure state updated successfully
...
  /orders:
    post:
      tags:
        - Order Service
      summary: Create a new order
      description: Initializes a new cart session with a unique order number and trace ID.
      servers:
        - url: http://localhost:8081
      parameters:
        - $ref: '#/components/parameters/HeaderStoreID'
        - $ref: '#/components/parameters/HeaderLaneID'
        - $ref: '#/components/parameters/HeaderLaneType'
        - $ref: '#/components/parameters/HeaderCashierID'
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '500':
          description: Database or server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/POSError'
  /orders/{orderNo}:
    get:
      tags:
        - Order Service
      summary: Get order details
      description: Retrieves the current state of an order including all items, subtotals, and taxes.
      servers:
        - url: http://localhost:8081
      parameters:
        - name: orderNo
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/HeaderStoreID'
        - $ref: '#/components/parameters/HeaderLaneID'
      responses:
        '200':
          description: Order details retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/POSError'
  /orders/{orderNo}/items:
    post:
      tags:
        - Order Service
      summary: Add item to order
      description: Adds a product to the cart by SKU. Validates active lane item limits from config.json.
      servers:
        - url: http://localhost:8081
      parameters:
        - name: orderNo
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/HeaderStoreID'
        - $ref: '#/components/parameters/HeaderLaneID'
        - $ref: '#/components/parameters/HeaderLaneType'
        - $ref: '#/components/parameters/HeaderCashierID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - sku:
              properties:
                sku:
                  type: string
                  example: "100001"
                quantity:
                  type: integer
                  example: 1
      responses:
        '200':
          description: Item added successfully
        '400':
          description: Item limit exceeded or order not OPEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/POSError'
        '404':
          description: SKU not found in catalog
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/POSError'
  /orders/{orderNo}/items/{sku}:
    delete:
      tags:
        - Order Service
      summary: Void item from cart
      description: Removes an item SKU from the active open order cart.
      servers:
        - url: http://localhost:8081
      parameters:
        - name: orderNo
          in: path
          required: true
          schema:
            type: string
        - name: sku
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/HeaderStoreID'
        - $ref: '#/components/parameters/HeaderLaneID'
      responses:
        '200':
          description: Item voided from cart
        '404':
          description: Item not found in cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/POSError'
  /orders/{orderNo}/calculate:
    post:
      tags:
        - Order Service
      summary: Calculate order total
      description: Computes subtotals, promotional discounts, taxes, and grand total.
      servers:
        - url: http://localhost:8081
      parameters:
        - name: orderNo
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/HeaderStoreID'
        - $ref: '#/components/parameters/HeaderLaneID'
      responses:
        '200':
          description: Calculation complete
          content:
            application/json:
              schema:
                type: object
                properties:
                  order_no:
                    type: string
                  trace_id:
                    type: string
                  subtotal:
                    type: number
                  discount:
                    type: number
                  tax:
                    type: number
                  total:
                    type: number
...
components:
  parameters:
    HeaderStoreID:
      name: X-Store-ID
      in: header
      required: false
      description: Store identifier (e.g. STORE-104)
      schema:
        type: string
        default: STORE-104
    HeaderLaneID:
      name: X-Lane-ID
      in: header
      required: false
      description: Checkout register lane ID (e.g. LANE-01, LANE-03)
      schema:
        type: string
        default: LANE-01
    HeaderLaneType:
      name: X-Lane-Type
      in: header
      required: false
      description: Checkout register lane type (CASHIER_EXPRESS, CASHIER_BELT, SELF_CHECKOUT)
      schema:
        type: string
        default: CASHIER_EXPRESS
    HeaderCashierID:
      name: X-Cashier-ID
      in: header
      required: false
      description: Operator / Cashier ID (e.g. CASHIER-101, CUSTOMER_SELF)
      schema:
        type: string
        default: CASHIER-101
  schemas:
    Order:
      type: object
      properties:
        id:
          type: integer
        order_no:
          type: string
          example: "ORD-10025"
        trace_id:
          type: string
          example: "TRC-82931"
        store_id:
          type: string
          example: "STORE-104"
        lane_id:
          type: string
          example: "LANE-03"
        lane_type:
          type: string
          example: "SELF_CHECKOUT"
        cashier_id:
          type: string
          example: "CUSTOMER_SELF"
        status:
          type: string
          example: "OPEN"
        subtotal:
          type: number
          format: float
          example: 45.00
        discount:
          type: number
          format: float
          example: 5.00
        tax:
          type: number
          format: float
          example: 3.40
        total:
          type: number
          format: float
          example: 43.40
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        sku:
          type: string
          example: "100001"
        name:
          type: string
          example: "Apples"
        category:
          type: string
          example: "Produce"
        quantity:
          type: integer
          example: 2
        unit_price:
          type: number
          format: float
          example: 4.99
        discount:
          type: number
          format: float
          example: 0.00
        tax:
          type: number
          format: float
          example: 0.85
        total:
          type: number
          format: float
          example: 10.83
...
    POSError:
      type: object
      properties:
        code:
          type: string
          example: "ERR_INSUFFICIENT_CASH"
        message:
          type: string
          example: "Insufficient cash received: Cash ($5.00) is less than Amount Due ($9.73)"
        category:
          type: string
          example: "BUSINESS"
...